% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_function.R
\name{ngram_function}
\alias{ngram_function}
\title{Function outputs top n-grams in a data frame}
\usage{
ngram_function(dataframe, ngram_number)
}
\arguments{
\item{dataframe}{input for the function}

\item{ngram_number}{number of n-grams desired (i.e 2 = bigram, 3 = trigram)}
}
\value{
data.frame() count of n-grams
}
\description{
Function outputs top n-grams in a data frame
}
\examples{
ngram_function(dataframe, 3)

}
\author{
Joshua Kim
}
