% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topwords_function.R
\name{topwords_function}
\alias{topwords_function}
\title{Function outputs a graph of top words}
\usage{
topwords_function(dataframe, top_n_number)
}
\arguments{
\item{dataframe}{input for the function}

\item{top_n_number}{number of words as output}
}
\value{
data.frame() graph of the top words in descending order
}
\description{
Function outputs a graph of top words
}
\examples{
topwords_function(dataframe, 30)

}
\author{
Joshua Kim
}
